refactor: update L1Withdrawer to pack and use smaller uints for gas limit and withdrawal amount

+++ packages/contracts-bedrock/interfaces/L2/IL1Withdrawer.sol
@@ -7,25 +7,25 @@ interface IL1Withdrawer is ISemver {
     error L1Withdrawer_OnlyProxyAdminOwner();
 
     event WithdrawalInitiated(uint256 amount, address indexed recipient);
-    event MinWithdrawalAmountUpdated(uint256 oldMinWithdrawalAmount, uint256 newMinWithdrawalAmount);
+    event MinWithdrawalAmountUpdated(uint216 oldMinWithdrawalAmount, uint216 newMinWithdrawalAmount);
     event RecipientUpdated(address oldRecipient, address newRecipient);
-    event WithdrawalGasLimitUpdated(uint256 oldWithdrawalGasLimit, uint256 newWithdrawalGasLimit);
+    event WithdrawalGasLimitUpdated(uint40 oldWithdrawalGasLimit, uint40 newWithdrawalGasLimit);
     event WithdrawalDataUpdated(bytes oldWithdrawalData, bytes newWithdrawalData);
 
-    function minWithdrawalAmount() external view returns (uint256);
+    function minWithdrawalAmount() external view returns (uint216);
     function recipient() external view returns (address);
-    function withdrawalGasLimit() external view returns (uint256);
+    function withdrawalGasLimit() external view returns (uint40);
     function withdrawalData() external view returns (bytes memory);
 
-    function setMinWithdrawalAmount(uint256 _newMinWithdrawalAmount) external;
+    function setMinWithdrawalAmount(uint216 _newMinWithdrawalAmount) external;
     function setRecipient(address _newRecipient) external;
-    function setWithdrawalGasLimit(uint256 _newWithdrawalGasLimit) external;
+    function setWithdrawalGasLimit(uint40 _newWithdrawalGasLimit) external;
     function setWithdrawalData(bytes memory _newWithdrawalData) external;
 
     function __constructor__(
-        uint256 _minWithdrawalAmount,
+        uint216 _minWithdrawalAmount,
         address _recipient,
-        uint256 _withdrawalGasLimit,
+        uint40 _withdrawalGasLimit,
         bytes memory _withdrawalData
     ) external;
 }
